paas-webinar-demo/requirements.txt
fastapi==0.133.0
uvicorn==0.40.0
openai==2.26.0
pydantic==2.11.7
python-dotenv==1.1.0
psycopg2-binary==2.9.10